Coded the operator Overloads in Array.hpp.
In the first it reallocates the memory size and then iterates over the array and sets values and then returns a pointer of "this".

For the next 2 [] operator overloads it checks to see if the index they have is valid and then returns the value stored in internalArray at the current index. The only difference between the 2 is that Type& is for the left hand operations.