Coded the testAdvancedArray() and added a default constructure to intialize.
The default constructor is in CrimeData.cpp which is then called in the public section in CrimeData.hpp.

Everything else in the controller was my own attempt to connect the files to the start so it can run and show up in the console.me We have to use an instance to call the methods from ArrayTester.