Added a protected section to the LinearNode.hpp file to fixe "next" errors, added a getFromIndex method, and moved a FileController method to the appropriate type of file.
This was all done to fix errors. The getFromIndex method was needed due to the fact there would be a VTable or reference error without it. The FileController methods were not ment to be in the header file. and the LinearNode<Type> * next; section in LinearNode.hpp is needed to even declare next and use it in other parts of the project like LinearNode.