Created and coded the SortingTester class and header file to do sorting for the Crime Data.
The SortingTester.cpp file contains sorting methods that can be used. many of them use the partision method in order to use pivots and such. Quicksort is the only type of sort in their now. It uses the pivot point and swap methods to read and select differnt values as needed and move them around. So far with the current code there is 77731 swaps done with the "index < 10000" section in the testSorts() section.