Did more coding for methods in BinarySearchTree such as the remove method and such.
The remove method takes a lot of code and check the different sides of the tree using multiple if, else if, and else statements to logically decide how to remove things and reconnect others. Then the get(right and left)MostChild methods are used to get the information fromt the corresponding child node.