Made a new header file in nodes called N_AryTreeNode.hpp.
Then coded the entire file with the class initializations, methods, constructors, getters & setters, etc.

This sort of structure/setup uses vectors and checks if there is a child in a section and if more "kids" can be added untill the limit is reached. There are no established left and right children. Once the limit is reached, it shift over to the next sections and repeats. Because of this, N_Ary Trees tend to be wide and possibly shallow.