Finished coding the GraphTester.cpp file which should've been done awhile ago.
The tester takes the items in the setup methods and uses the various methods in the Graph template file to go through and get information. since setup() and traverseGraph() are private methods, a public method is used to call them so they may be used. NOTE: The tester isn't working quite right for the Compare Graph Traversals assignment

Also attempted to correctly make a UML diagram that involves List, LinkedList, Stack, and Queue. It may not be correct.